* 'master' of https://gitee.com/spruceovo/arkui_ace_engine: (106 commits)
  add
  add ai event report
  补充OnWillChange的实例代码 Signed-off-by: haoshuo <haoshuo1@huawei.com>
  [Repeat]代码同步pu_repeat_virtual_scroll_2_impl.ts(2)
  ui_context新增innerapi接口
  res
  修复主动拖拽跟手图跳变的问题
  uec visible
  优化CheckFilterProperty
  add TDD coverage
  a
  [Repeat]代码同步pu_repeat_virtual_scroll_2_impl.ts
  添加 tdd，提高 EstimateTotalHeight 覆盖率
  updateBuilder
  有符号转无符号
  Eliminate the compilation error
  fix caret avoid not avoid
  全局坐标系——小数差异问题修复
  Revert "add tdd"
  fixMenuErr
  ...

+++ frameworks/core/components_ng/render/adapter/rosen_window.cpp
@@ -200,7 +200,7 @@ void RosenWindow::PostVsyncTimeoutDFXTask(const RefPtr<TaskExecutor>& taskExecut
         LOGW("ArkUI request vsync, but no vsync received in 500ms");
         auto window = weakWindow.lock();
         if (window) {
-            uint64_t nanoTimestamp = GetSysTimestamp();
+            uint64_t nanoTimestamp = static_cast<uint64_t>(GetSysTimestamp());
             // force flush vsync with now time stamp and UINT64_MAX as frameCount.
             window->ForceFlushVsync(nanoTimestamp, UINT64_MAX);
         }
